producer/requirements.txt
black==21.12b0
boto3==1.20.42
boto3-stubs==1.20.42
botocore==1.23.42
botocore-stubs==1.23.42
click==8.0.3
gevent==21.12.0
greenlet==1.1.2
isort==5.10.1
jmespath==0.10.0
mypy==0.931
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pydantic==1.9.0
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.5.0
six==1.16.0
tomli==1.2.3
types-PyYAML==6.0.3
typing_extensions==4.0.1
urllib3==1.26.8
websocket-client==1.2.3
zope.event==4.5.0
zope.interface==5.4.0
